to improve TrackViewHolder code (exl. inner class)
