+++ app/src/main/java/ru/startandroid/develop/sprint8v3/TrackViewHolder.kt
@@ -8,15 +8,18 @@ import com.bumptech.glide.Glide
 import com.bumptech.glide.request.RequestOptions
 
 class TrackViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
-    private val trackNameTextView: TextView
-    private val artistNameTextView: TextView
-    private val trackTimeTextView: TextView
-    private val artworkImageView: ImageView
+    private val trackNameTextView: TextView = itemView.findViewById(R.id.trackNameTextView)
+    private val artistNameTextView: TextView = itemView.findViewById(R.id.artistNameTextView)
+    private val trackTimeTextView: TextView = itemView.findViewById(R.id.trackTimeTextView)
+    private val artworkImageView: ImageView = itemView.findViewById(R.id.artworkImageView)
 
-    init {
-        trackNameTextView = itemView.findViewById(R.id.trackNameTextView)
-        artistNameTextView = itemView.findViewById(R.id.artistNameTextView)
-        trackTimeTextView = itemView.findViewById(R.id.trackTimeTextView)
-        artworkImageView = itemView.findViewById(R.id.artworkImageView)
+    fun bind(model: Track) {
+        trackNameTextView.text = model.trackName
+        artistNameTextView.text = model.artistName
+        trackTimeTextView.text = model.trackTime
+        Glide.with(itemView.context)
+            .load(model.artworkUrl100)
+            .apply(RequestOptions().placeholder(R.drawable.placeholder_image))
+            .into(artworkImageView)
     }
 }
